chore: move non-cobra methods out of cmd
