+++ internal/cmd/flags.go
@@ -2,13 +2,10 @@ package cmd
 
 import (
 	"fmt"
-	"time"
 
-	"github.com/ccoVeille/gh-reaction/internal/timeago"
+	"github.com/ccoVeille/gh-reaction/internal/app"
 	"github.com/spf13/cobra"
 )
-
-const defaultSinceDaysAgo = 90
 
 // addSinceFlagToCmd adds the since flag to a command to avoid duplication
 func addSinceFlagToCmd(cmd *cobra.Command) {
@@ -16,22 +13,6 @@ func addSinceFlagToCmd(cmd *cobra.Command) {
 		&sinceFlag,
 		"since",
 		"",
-		fmt.Sprintf(`Fetch messages since this date (e.g., "2023-01-02", "2h", "15m", "3d" ...) (default "%dd")`, defaultSinceDaysAgo),
+		fmt.Sprintf(`Fetch messages since this date (e.g., "2023-01-02", "2h", "15m", "3d" ...) (default "%dd")`, app.DefaultSinceDaysAgo),
 	)
-}
-
-// ParseSinceFlag parses the since flag and returns a RelativeDate
-func ParseSinceFlag() (timeago.RelativeDate, error) {
-	var since timeago.RelativeDate
-
-	if sinceFlag == "" {
-		since = timeago.NewRelativeDate(time.Now().AddDate(0, 0, -defaultSinceDaysAgo))
-	} else {
-		if err := since.Set(sinceFlag); err != nil {
-			return since, err
-		}
-	}
-
-	since.Time = since.Time.Truncate(time.Hour).UTC()
-	return since, nil
 }
